refactor(data): move instrumentation into services
