+++ src/renderer/data/utils/dataApiDevtools.ts
@@ -1,7 +1,7 @@
 import { isDev } from '@renderer/utils/platform'
 import type { DataResponse, HttpMethod } from '@shared/data/api/types'
 
-import { dataApiInstrumentationService } from './DataApiInstrumentationService'
+import { dataApiInstrumentationService } from '../services/DataApiInstrumentationService'
 
 type DataApiDevtoolsRequestState = 'pending' | 'success' | 'error' | 'retry'
